Updates to the Language Guide: Haskell page
The Language Guide: Haskell page has been updated to use
epoch when creating a cache, in order to work around potential
issues caused by ignoring the version of stack used in the docker
image and by ignoring the resolver used in stack.yaml.

Lastly, the sample configuration at the top of the page was
updated to cache ".stack", in addition to ".stack-work".