Explicit about need for Jira Cloud for orb support
The orb integration to report build statuses depends on JWT tokens and Atlassian Connect, which is only available on their cloud offering. This is outside the control of CircleCI.

This change updates our docs to be explicit about this difference. Other features like "create an issue" will still work on server.